requirements.txt
# Core dependencies for Ultimate Agent Directory build system
pyyaml>=6.0.1          # YAML parsing
pydantic>=2.0.0        # Schema validation
jinja2>=3.1.0          # Template rendering

# Optional (for future features)
requests>=2.31.0       # For GitHub API
python-frontmatter>=1.0.0  # If we want YAML frontmatter in markdown
aiohttp>=3.9.0         # Async HTTP for link checker
tqdm>=4.66.0           # Progress bars for link checker

# Development/Testing
pytest>=8.0.0          # Test framework
ruff>=0.14.11          # Linter
mypy>=1.19.1           # Type checker
pytest-cov>=4.0.0      # Coverage reporting
